src/contacts/requirements.txt
flask
pyjwt
cryptography
